[JENKINS-57903]- Refactor PromotionConditionDescriptor class and other dependent files
