+++ src/main/java/hudson/plugins/promoted_builds/PromotionConditionDescriptor.java
@@ -2,6 +2,10 @@ package hudson.plugins.promoted_builds;
 
 import hudson.model.AbstractProject;
 import hudson.model.Descriptor;
+import hudson.model.Job;
+import hudson.model.TaskListener;
+
+import javax.annotation.Nonnull;
 
 /**
  * {@link Descriptor} for {@link PromotionCondition}.
@@ -25,5 +29,11 @@ public abstract class PromotionConditionDescriptor extends Descriptor<PromotionC
      * @return
      *      true to allow user to configure this promotion condition for the given project.
      */
+    public boolean isApplicable(@Nonnull Job<?,?> item,@Nonnull TaskListener listener){
+        if(item instanceof AbstractProject){
+            return isApplicable((AbstractProject)item, TaskListener.NULL);
+        }
+        return isApplicable((Job) item, listener);
+    }
     public abstract boolean isApplicable(AbstractProject<?,?> item);
 }
