+++ src/main/java/hudson/plugins/promoted_builds/conditions/GroovyConditionDescriptor.java
@@ -4,9 +4,13 @@ import hudson.Extension;
 import hudson.PluginManager;
 import hudson.PluginWrapper;
 import hudson.model.AbstractProject;
+import hudson.model.Job;
+import hudson.model.TaskListener;
 import hudson.plugins.promoted_builds.PromotionConditionDescriptor;
+import javafx.concurrent.Task;
 import jenkins.model.Jenkins;
 
+import javax.annotation.Nonnull;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -14,7 +18,7 @@ import java.util.logging.Logger;
  * Declared outside GroovyCondition as it depends on the optional 'script-security' dependency
  */
 @Extension(optional = true)
-public final class GroovyConditionDescriptor extends PromotionConditionDescriptor {
+public class GroovyConditionDescriptor extends PromotionConditionDescriptor {
 
     private static final Logger LOGGER = Logger.getLogger(GroovyConditionDescriptor.class.getName());
 
@@ -22,22 +26,30 @@ public final class GroovyConditionDescriptor extends PromotionConditionDescripto
         super(GroovyCondition.class);
     }
 
-    @Override
-    public boolean isApplicable(final AbstractProject<?, ?> item) {
-        // TODO switch to Jenkins.getActiveInstance() once bumped to 1.590
-        final Jenkins jenkins = Jenkins.getInstance();
-        if (jenkins == null) {
-            // Jenkins not started or shut down
-            return false;
+   public boolean isApplicable(@Nonnull Job<?,?> item, TaskListener listener){
+        if(item instanceof AbstractProject){
+            return isApplicable((AbstractProject)item, TaskListener.NULL);
+        }else{
+            final Jenkins jenkins = Jenkins.getInstance();
+            if (jenkins == null) {
+                // Jenkins not started or shut down
+                return false;
+            }
+            final PluginManager pluginManager = jenkins.getPluginManager();
+            if (pluginManager == null) {
+                LOGGER.log(Level.WARNING, "No PluginManager");
+                return false;
+            }
+            final PluginWrapper plugin = pluginManager.getPlugin("script-security");
+            return plugin != null && plugin.isActive();
         }
-        final PluginManager pluginManager = jenkins.getPluginManager();
-        if (pluginManager == null) {
-            LOGGER.log(Level.WARNING, "No PluginManager");
-            return false;
-        }
-        final PluginWrapper plugin = pluginManager.getPlugin("script-security");
-        return plugin != null && plugin.isActive();
+
+   }
+
+   public boolean isApplicable(AbstractProject<?, ?> item) {
+        return false;
     }
+
 
     @Override
     public String getDisplayName() {
